moving the multi-segment related to MSW directory.
