+++ opm/parser/eclipse/CMakeLists.txt
@@ -94,9 +94,9 @@ EclipseState/Schedule/Well.cpp
 EclipseState/Schedule/WellProductionProperties.cpp
 EclipseState/Schedule/WellInjectionProperties.cpp
 EclipseState/Schedule/WellPolymerProperties.cpp
-EclipseState/Schedule/Segment.cpp
-EclipseState/Schedule/SegmentSet.cpp
-EclipseState/Schedule/Compsegs.cpp
+EclipseState/Schedule/MSW/Segment.cpp
+EclipseState/Schedule/MSW/SegmentSet.cpp
+EclipseState/Schedule/MSW/Compsegs.cpp
 EclipseState/Schedule/WellSet.cpp
 EclipseState/Schedule/Group.cpp
 EclipseState/Schedule/Completion.cpp
@@ -190,9 +190,9 @@ EclipseState/Schedule/Well.hpp
 EclipseState/Schedule/WellProductionProperties.hpp
 EclipseState/Schedule/WellInjectionProperties.hpp
 EclipseState/Schedule/WellPolymerProperties.hpp
-EclipseState/Schedule/Segment.hpp
-EclipseState/Schedule/SegmentSet.hpp
-EclipseState/Schedule/Compsegs.hpp
+EclipseState/Schedule/MSW/Segment.hpp
+EclipseState/Schedule/MSW/SegmentSet.hpp
+EclipseState/Schedule/MSW/Compsegs.hpp
 EclipseState/Schedule/WellSet.hpp
 EclipseState/Schedule/Group.hpp
 EclipseState/Schedule/DynamicState.hpp
