+++ opm/parser/eclipse/EclipseState/Schedule/MSW/Segment.cpp
@@ -1,4 +1,4 @@
-#include <opm/parser/eclipse/EclipseState/Schedule/Segment.hpp>
+#include <opm/parser/eclipse/EclipseState/Schedule/MSW/Segment.hpp>
 
 
 namespace Opm {
